refactor/public response
